Fixed Code smell, and added logger for device info and stream request (#39)
Code Fix for Sonar Code smell

+++ registration/registration-services/src/main/java/io/mosip/registration/service/bio/impl/BioServiceImpl.java
@@ -304,7 +304,7 @@ public class BioServiceImpl extends BaseService implements BioService {
 		}
 
 		setBioQualityScores(requestDetail.getType(), attempt,
-				currentCaptureQualityScore / mosipBioDeviceDataResponses.size());
+				currentCaptureQualityScore / (double)mosipBioDeviceDataResponses.size());
 
 		double slapQuality = (fpDetailsDTO.getSegmentedFingerprints().stream()
 				.mapToDouble(finger -> finger.getQualityScore()).sum())
